# api/openapi.yaml
openapi: 3.1.0
info:
  title: sing-box WebUI Local API
  version: 0.1.0
  description: Local loopback API for the sing-box WebUI control plane.
servers:
  - url: http://127.0.0.1:11872
paths:
...
  /api/v1/subscriptions/{id}/refresh:
    parameters:
      - $ref: '#/components/parameters/SubscriptionID'
      - $ref: '#/components/parameters/CSRFToken'
    post:
      operationId: refreshSubscription
      responses:
        '200':
          description: Refreshed subscription.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/subscriptions/{id}/activate:
    parameters:
      - $ref: '#/components/parameters/SubscriptionID'
      - $ref: '#/components/parameters/CSRFToken'
    post:
      operationId: activateSubscription
      responses:
        '200':
          description: Active subscription changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/subscriptions/{id}/selection:
    parameters:
      - $ref: '#/components/parameters/SubscriptionID'
      - $ref: '#/components/parameters/CSRFToken'
    put:
      operationId: selectNode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [nodeId]
              properties:
                nodeId:
                  type: string
      responses:
        '200':
          description: Selected node changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/subscriptions/{id}/latency:
    parameters:
      - $ref: '#/components/parameters/SubscriptionID'
      - $ref: '#/components/parameters/CSRFToken'
    post:
      operationId: testNodeLatency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LatencyRequest'
      responses:
        '200':
          description: End-to-end HTTPS delay results measured through temporary sing-box outbounds.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LatencyResponse'
        default:
          $ref: '#/components/responses/Error'
...
components:
  parameters:
    SubscriptionID:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Structured API error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Node:
      type: object
      additionalProperties: false
      required: [id, name, type, server, port, tls, selected]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        server:
          type: string
        port:
          type: integer
          minimum: 1
          maximum: 65535
        tls:
          type: boolean
        selected:
          type: boolean
    Subscription:
      type: object
      additionalProperties: false
      required: [id, name, url, autoUpdate, updateIntervalMinutes, active, nodeCount]
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
        autoUpdate:
          type: boolean
        updateIntervalMinutes:
          type: integer
        active:
          type: boolean
        selectedNodeId:
          type: string
        lastUpdated:
          type: string
          format: date-time
        lastError:
          type: string
        lastFetchPath:
          type: string
          enum: [direct, proxy]
          description: 上次成功拉取所用的路径（direct 直连 / proxy 代理回退）。
        nodeCount:
          type: integer
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/Node'
...
    LatencyRequest:
      type: object
      additionalProperties: false
      properties:
        nodeIds:
          type: array
          maxItems: 200
          uniqueItems: true
          items:
            type: string
...
    LatencyResult:
      type: object
      additionalProperties: false
      required: [nodeId, name, status]
      properties:
        nodeId:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [ok, timeout, failed]
        latencyMs:
          type: integer
          minimum: 1
        detail:
          type: string
    LatencyResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LatencyResult'
...
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, requestId]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
        requestId:
          type: string